update docker to don't reload dependencies at each start WARNING Bug with app.py and init.py of backend python

+++ docker/front_DashPlotly/entrypoint.sh
@@ -1,10 +1,20 @@
 #!/bin/sh
+# Be careful don't use bash and bash command
 
-python -m venv venv
+cd /app
 
-source venv/bin/activate
+python -m venv docker_venv
 
-pip install --upgrade pip
-pip install -r requirements.txt
+. docker_venv/bin/activate
+
+# Vérifier si l'installation a déjà été effectuée
+if [ ! -f "docker_venv/.installed" ]; then
+    echo "Première installation des dépendances..."
+    pip install --upgrade pip
+    pip install -r requirements.txt
+    touch docker_venv/.installed
+else
+    echo "Dépendances déjà installées, on passe..."
+fi
 
 python main.py
